\name{initialize-methods}
\docType{methods}
\alias{initialize,ProjectInfo-method}
\alias{initialize,TreeSet-method}
\alias{initialize,SchemeTreeSet-method}
\alias{initialize,ProcesSet-method}
\alias{initialize,DataTreeSet-method}
\alias{initialize,ExprTreeSet-method}
\alias{initialize,CallTreeSet-method}
\alias{initialize,QualTreeSet-method}
\alias{initialize,Filter-method}
\alias{initialize,PreFilter-method}
\alias{initialize,UniFilter-method}
\alias{initialize,FilterTreeSet-method}
\alias{initialize,AnalysisTreeSet-method}
\alias{initialize-methods}
\alias{initialize}
\title{Initialize Classes}
\description{
 Initialize S4 classes.
}
\section{Methods}{
\describe{
 Internal method to initialize S4 classes.
}
}
\keyword{methods}
