\name{xps-package}
\alias{xps-package}
\alias{xps}
\docType{package}
\title{xps Package Overview}
\description{
   xps Package Overview
}
\details{
  Important data classes: \code{\link{SchemeTreeSet}},
  \code{\link{DataTreeSet}}
  \code{\link{ExprTreeSet}}
  \code{\link{CallTreeSet}}. Full help on methods and
  associated functions is available from within class help pages.

  Additional data classes: \code{\link{ProjectInfo}}.

  The package handles the pre-processing and normalization 
  of Affymetrix GeneChip expression arrays, including GeneChip exon
  array systems (Exon 1.0 ST) and gene array systems (Gene 1.0 ST). 
  It is an R wrapper to XPS (eXpression Profiling System), which 
  is based on ROOT, an object-oriented framework developed at CERN. 
  Thus, the prior installation of ROOT is a prerequisite for the 
  usage of this package, see the README file. ROOT is licensed 
  under LGPL and can be downloaded from <http://root.cern.ch>.
}
\author{
   Christian Stratowa \email{cstrato@aon.at}
}
\keyword{package}
