\name{xpsOptions}
\alias{xpsOptions}
\alias{debug.xps}
\title{xps Options}
\description{
  Options for xps
}
\usage{
xpsOptions(debug=FALSE)
}
\arguments{
  \item{debug}{logical, if \code{TRUE}, print debug information.}
}
\details{
Currently only used to set \code{debug} to \code{FALSE} or \code{TRUE}.
}
\value{
A global variable \code{debug.xps} can be set to \code{TRUE}.
}
\author{Christian Stratowa}
\keyword{manip}
