% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{xps2df}
\alias{xps2df}
\title{Import data from XPS-ESCA spectrometer}
\usage{
xps2df(xps.path)
}
\arguments{
\item{xps.path}{Path to file with XPS data in expected format}
}
\value{
dataframe with
    + interval : chr
    + xstart   : num
    + xstep    : num
    + x        : num
    + y        : num
}
\description{
Import ASCII text-file with one or more intervals of XPS data,
where each interval (transition) specifies a description,
x-start (binding energy), x-step, and number of rows.
These ASCII files commonly use the *.asc extension, but may use
other extensions.
}
